+++ lib/rqrcode_core/qrcode/qr_alphanumeric.rb
@@ -1,7 +1,11 @@
 # frozen_string_literal: true
 
 module RQRCodeCore
-  ALPHANUMERIC = ["0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z", " ", "$", "%", "*", "+", "-", ".", "/", ":"]
+  ALPHANUMERIC = [
+    "0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "A", "B", "C", "D", "E", "F", "G", "H", "I",
+    "J", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z", " ", "$",
+    "%", "*", "+", "-", ".", "/", ":"
+  ].freeze
 
   class QRAlphanumeric
     attr_reader :mode