+++ lib/rqrcode_core/qrcode/qr_numeric.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module RQRCodeCore
-  NUMERIC = ["0", "1", "2", "3", "4", "5", "6", "7", "8", "9"].freeze
+  NUMERIC = %w[0 1 2 3 4 5 6 7 8 9].freeze
 
   class QRNumeric
     attr_reader :mode